fix backend

+++ backend/src/run_game_data.rs
@@ -3,16 +3,16 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug, Default, Apiv2Schema)]
 pub struct RunGameData {
-    #[serde(default, rename = "Bot1")]
+    #[serde(rename = "Bot1")]
     pub bot1: Vec<String>,
-    #[serde(default, rename = "Bot2")]
+    #[serde(rename = "Bot2")]
     pub bot2: Vec<String>,
-    #[serde(default, rename = "Map")]
+    #[serde(rename = "Map")]
     pub map: Vec<String>,
-    #[serde(default, rename = "Iterations")]
-    pub iterations: u64,
-    #[serde(default, rename = "Visualize")]
+    #[serde(rename = "Iterations")]
+    pub iterations: i32,
+    #[serde(rename = "Visualize")]
     pub visualize: bool,
-    #[serde(default, rename = "Realtime")]
+    #[serde(rename = "Realtime")]
     pub realtime: bool,
 }
